requirements.txt
marker-pdf
transformers
torch
sentencepiece
protobuf